requirements.txt
insightface
opencv-python
streamlit
pandas
sqlalchemy
pillow
torch
plotly
